requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
google-generativeai==0.3.2
pinecone-client==3.0.0
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
streamlit==1.28.0
